ansible_playbook_from_stig/requirements.txt
# STIG to Ansible Playbook Generator Dependencies

# Core dependencies
requests>=2.28.0          # For LLM API calls
PyYAML>=6.0               # For Ansible playbook generation and validation (REQUIRED)
python-dotenv>=1.0.0      # For .env file loading

# LangGraph and LangChain dependencies
langgraph==0.5.2          # For workflow orchestration
langchain==0.3.26         # For LLM abstraction
langchain-community==0.3.27 # For community integrations
aiohttp==3.12.14          # For async HTTP requests

# Optional enhancements
click>=8.0.0              # For enhanced CLI interface
jinja2>=3.0.0             # For playbook templating

# Development dependencies (optional)
pytest>=7.0.0             # For testing
black>=22.0.0             # For code formatting

# Note: PyYAML is now REQUIRED for proper YAML validation and structure checking
# Install with: pip install PyYAML
